Fix bug 4650547: Az.ResourceGraph module is not imported before Graph query (#61)
If CheckPlannedMaintenanceNotification is set to true, the AzureSqlBulkFailover runbook tries to query Azure Resource Graph for planned maintenance notification information. This fails with "The 'Search-AzGraph' command was found in the module 'Az.ResourceGraph', but the module could not be loaded. For more information, run 'Import-Module Az.ResourceGraph'."

The simple solution would be "Import-Module Az.ResourceGraph", but this is blocked by "This module requires Az.Accounts version 4.2.0. An earlier version of Az.Accounts is imported in the current PowerShell session. Please open a new session before importing this module. This error could indicate that multiple incompatible versions of the Azure PowerShell cmdlets are installed on your system. Please see https://aka.ms/azps-version-error for troubleshooting information."

The latter error occurs because Azure Automation imports an older 2.* version of Az.Accounts, and the latest version of Az.ResourceGraph is not compatible with this version.

To sidestep the module version compat issues, we switch to direct REST calls to Graph. Test details in 4650547.

Testing performed by deploying a runbook from private branch. Scenarios:

End-to-end successful execution, including UpgradeMeNow/Failover calls.
GetPlannedNotificationId() can retrieve a planned maintenance event ID after the switch to Invoke-AzRestMethod.
Runbook execution fails with the expected message if there is no pending CMW maintenance notification and the CheckPlannedMaintenanceNotification runbook account variable is set to True.